requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
scipy>=1.7.0
flask>=2.0.1
flask-cors>=3.0.10
joblib>=1.0.1
nltk>=3.6.2
torch>=1.9.0
transformers>=4.9.0
python-dotenv>=0.19.0
requests>=2.26.0
tqdm>=4.62.0
optuna>=2.10.0 